requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
discord.py==2.3.2
executing==2.0.0
fonttools==4.42.1
frozendict==2.3.8
frozenlist==1.4.0
html5lib==1.1
icecream==2.1.3
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.25.2
packaging==23.1
pandas==2.1.0
peewee==3.16.3
Pillow==10.0.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.4
webencodings==0.5.1
yarl==1.9.2
yfinance==0.2.30
